requirements.txt
numpy
pyyaml
torch
tqdm
rasterio
scipy
matplotlib
earthengine-api
geopandas
shapely
pyproj
